feat(install): enhance installation script with feedback
Added feedback messages for various installation steps including yay,
packages, git configuration, oh-my-zsh, zsh plugins, miniconda3, nvm,
dotfiles, themes, and obs-studio. Improved cleanup process by removing
unnecessary packages and added checks before copying themes.

+++ install.sh
@@ -3,16 +3,40 @@
 echo "Starting installation..."
 
 # Install yay
-echo "Installing yay..."
-sudo pacman -S --needed git base-devel
-git clone https://aur.archlinux.org/yay.git ~/yay
-cd ~/yay
-makepkg -si
-rm -rf ~/yay
+if ! command -v yay &> /dev/null; then
+  echo "Installing yay..."
+
+  sudo pacman -S --needed git base-devel
+  git clone https://aur.archlinux.org/yay.git ~/yay
+  cd ~/yay
+  makepkg -si
+  rm -rf ~/yay
+
+  echo "yay installed!"
+else
+  echo "yay is installed!"
+fi
 
 # Install packages
 echo "Installing packages..."
-yay -S zsh hyprpaper hyprpicker hypridle hyprlock wlogout wofi pavucontrol brightnessctl playerctl floorp-bin cliphist wl-clipboard grim slurp thunar gvfs lsd bat nwg-look p7zip fastfetch btop commitizen-go localsend lazygit noto-fonts noto-fonts-cjk noto-fonts-emoji github-cli
+yay -S zsh hyprpaper hyprpicker hypridle hyprlock wlogout wofi pavucontrol brightnessctl playerctl floorp-bin cliphist wl-clipboard grim slurp thunar gvfs lsd bat nwg-look p7zip fastfetch btop commitizen-go localsend lazygit noto-fonts noto-fonts-cjk noto-fonts-emoji github-cli docker
+echo "Packages installed!"
+
+# git config
+echo "Configuring git..."
+
+git config --global core.editor "nvim"
+if [ -z "$(git config --global user.name)" ]; then
+  read -p "Enter your name: " name
+  git config --global user.name "$name"
+fi
+
+if [ -z "$(git config --global user.email)" ]; then
+  read -p "Enter your email: " email
+  git config --global user.email "$email"
+fi
+
+echo "Git config completed!"
 
 # Generate ssh key and add to github
 echo "Generating ssh key..."
@@ -27,59 +51,125 @@ else
 fi
 
 # Install oh-my-zsh
-echo "Installing oh-my-zsh..."
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+if [ ! -d ~/.oh-my-zsh ]; then
+  echo "Installing oh-my-zsh..."
+  sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+  echo "oh-my-zsh installed!"
+fi
 
 # Install zsh plugins
 echo "Installing zsh plugins..."
-git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
-git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
+
+if [ ! -d ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions ]; then
+  git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
+fi
+
+if [ ! -d ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting ]; then
+  git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
+fi
+
+echo "Zsh plugins installed!"
 
 # Install miniconda3
-read -p "Do you want to install miniconda3? [y/n]: " is_miniconda
-if [ $is_miniconda == "y" ]; then
-  echo "Installing miniconda3..."
-  mkdir -p ~/.miniconda3
-  wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/.miniconda3/miniconda.sh
-  bash ~/.miniconda3/miniconda.sh -b -u -p ~/.miniconda3
-  rm -rf ~/.miniconda3/miniconda.sh
-  ~/.miniconda3/bin/conda init zsh
+if [ ! -d ~/.miniconda3 ]; then
+  read -p "Do you want to install miniconda3? [y/n]: " is_miniconda
+  if [ $is_miniconda == "y" ]; then
+    echo "Installing miniconda3..."
+
+    mkdir -p ~/.miniconda3
+    wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/.miniconda3/miniconda.sh
+    bash ~/.miniconda3/miniconda.sh -b -u -p ~/.miniconda3
+    rm -rf ~/.miniconda3/miniconda.sh
+    ~/.miniconda3/bin/conda init zsh
+
+    echo "Miniconda3 installed!"
+  fi
 fi
 
 # Install nvm
-read -p "Do you want to install nvm? [y/n]: " is_nvm
-if [ $is_nvm == "y" ]; then
-  echo "Installing nvm..."
-  curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.0/install.sh | bash
+if [ ! -d ~/.nvm ]; then
+  read -p "Do you want to install nvm? [y/n]: " is_nvm
+  if [ $is_nvm == "y" ]; then
+    echo "Installing nvm..."
+    curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.0/install.sh | bash
+    echo "nvm installed!"
+  fi
 fi
 
 # Clone dotfiles
 echo "Cloning dotfiles..."
-git clone git@github.com:tiesen243/dotfiles.git ~/dotfiles
+
+if [ ! -d ~/dotfiles ]; then
+  git clone git@github.com:tiesen243/dotfiles.git ~/dotfiles
+fi
+
+echo "Dotfiles cloned!"
 
 # Linking dotfiles to .config directory
 echo "Linking dotfiles..."
+
+rm ~/.zshrc
+rm -rf ~/.config/{hypr,btop,kitty,dunst,wlogout,lazygit,fastfetch}
+rm -rf ~/.oh-my-zsh/custom/themes/yuki.zsh-theme
+
 ln -s ~/dotfiles/hypr ~/.config
 ln -s ~/dotfiles/btop ~/.config
 ln -s ~/dotfiles/kitty ~/.config
 ln -s ~/dotfiles/dunst ~/.config
+ln -s ~/dotfiles/wlogout ~/.config
 ln -s ~/dotfiles/lazygit ~/.config
 ln -s ~/dotfiles/fastfetch ~/.config
 ln -s ~/dotfiles/zsh/config.zsh ~/.zshrc
 ln -s ~/dotfiles/zsh/themes/yuki.zsh-theme ~/.oh-my-zsh/custom/themes
 
+echo "Dotfiles linked!"
+
 # Copying themes
 echo "Copying themes..."
-cp -a ~/dotfiles/themes/. ~/
+
+if [ ! -d ~/.themes ]; then
+  cp -r ~/dotfiles/themes/.themes ~/
+fi
+
+if [ ! -d ~/.icons ]; then
+  cp -r ~/dotfiles/themes/.icons ~/
+fi
+
+if [ ! -d ~/dotfiles/themes/.fonts ]; then
+  cp -r ~/dotfiles/themes/.fonts ~/
+fi
+
 echo "Themes copied! You can change the theme in 'nwg-look'."
 
 # Install obs-studio
-read -p "Do you want to install obs-studio? [y/n]: " is_obs
-if [ $is_obs == "y" ]; then
-  yay -S obs-studio wlrobs-hg
+if ! command -v obs &> /dev/null; then
+  read -p "Do you want to install obs-studio? [y/n]: " is_obs
+  if [ $is_obs == "y" ]; then
+    echo "Installing obs-studio..."
+    yay -S obs-studio wlrobs-hg
+    echo "OBS Studio installed!"
+  fi
 fi
 
 echo "Cleaning up..."
-yay -Qdtq | yay -Rns -
+
+if [ -n "$(yay -Qdtq)" ]; then
+  yay -Runs $(yay -Qdtq)
+fi
+
+# remove unnecessary packages
+if command -v vim &> /dev/null; then
+  sudo pacman -Runs vim
+fi
+
+if command -v nano &> /dev/null; then
+  sudo pacman -Runs nano
+fi 
+
+if command -v dolphin &> /dev/null; then
+  sudo pacman -Runs dolphin
+fi
+
+echo "Cleaning up completed!"
 
 echo "Installation completed!"
